pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudio"
version = "0.1.0"
description = "Easily orchestrate automations using CloudIO"
authors = [
    {name = "CloudIO Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow==2.9.3"
]